Split X-specific code into x.c
