+++ Makefile
@@ -3,7 +3,7 @@
 
 include config.mk
 
-SRC = st.c
+SRC = st.c x.c
 OBJ = ${SRC:.c=.o}
 
 all: options st
@@ -20,6 +20,9 @@ config.h:
 .c.o:
 	@echo CC $<
 	@${CC} -c ${CFLAGS} $<
+
+st.o: config.h st.h win.h
+x.o: arg.h st.h win.h
 
 ${OBJ}: config.h config.mk
